# Disable typechecking on the name attribute for TypedObject.